REFACTOR: 파일이름 및 폴더구조 변경
