+++ FE-Project/src/page/Home.tsx
@@ -1,16 +1,16 @@
 import React from "react";
 import Footer from "../components/footer/Footer";
 import Header from "../components/header/Header";
-import ProjectCardlist from "../components/home/ProjectCardlist";
 import MainCarousel from "../components/home/MainCarousel";
-import PromotionCardlist from "../components/home/PromotionCardlist";
+import MainRecruitmentCardlist from "../components/home/MainRecruitmentCardlist";
+import MainPromotionCardlist from "../components/home/MainPromotionCardlist";
 export default function Home() {
   return (
     <>
       <Header />
       <MainCarousel />
-      <ProjectCardlist />
-      <PromotionCardlist />
+      <MainRecruitmentCardlist />
+      <MainPromotionCardlist />
       <Footer />
     </>
   );
